requirements.txt
flask==2.3.3
google-generativeai==0.3.1
fpdf==1.7.2
python-dotenv==1.0.0